Refactor cuda_utils : move custatevec to lightning_gpu_utils (#681)
* Init commit

* Auto update version

* Add MidMeasureMP Kokkos (#658)

* collapse() and normalized() with Kokkos

* Auto update version

* CodeFactor

* normalize() abort if norm is zero

* unittest for normalize()

* Auto update version

* unittest for normalize() in LighningQubit

* Auto update version

* Add MCM bindings and tests for L-Kokkos. (#672)

* Add MCM bindings and tests for L-Kokkos.

* Auto update version

* Refactor collapse.

* Try parallelizing tests.

* Only run test_native_mcm once amongst all workflows since it takes quite a bit of time.

* trigger ci

---------


* Do not run test_native_mcm with LK-GPU.

* Auto update version

* trigger ci

* Update pennylane_lightning/core/src/simulators/lightning_kokkos/bindings/LKokkosBindings.hpp


* Auto update version

* Update pennylane_lightning/core/src/simulators/lightning_kokkos/StateVectorKokkos.hpp


* Update pennylane_lightning/core/src/simulators/lightning_kokkos/StateVectorKokkos.hpp


* Update pennylane_lightning/core/src/simulators/lightning_kokkos/StateVectorKokkos.hpp


* Update pennylane_lightning/core/src/simulators/lightning_kokkos/StateVectorKokkos.hpp


* Increase shotsfor flaky test_composite_mcm_single_measure_obs

* Create separate device_allowed_operations list for ops supported by the device but not Catalyst.

* Use isinstance instead of name comparison.

* Check projector.

* Update operations at device init.

* Modify stopping condition in LK.

* Cannot use instance with Hamiltonian check.

* Update pennylane_lightning/core/src/simulators/lightning_kokkos/tests/Test_StateVectorLKokkos.cpp


* Update pennylane_lightning/core/src/simulators/lightning_kokkos/StateVectorKokkos.hpp


* Remove obsolete comment.

* Auto update version

* trigger ci

---------


* Trigger CI

* make format

* Auto update version

* Trigger CIs

* Auto update version

* Trigger CI

* Auto update version

* add more docstring

* add changelog

* Update pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVecError.hpp


* Update pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVecError.hpp


* Update pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVec_helpers.hpp


* Update pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVecError.hpp


* Update pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVec_helpers.hpp


* Update pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVec_helpers.hpp


* Update pennylane_lightning/core/src/utils/cuda_utils/cuError.hpp


* Update pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVecError.hpp


* Trigger CI

* make format

* quick fix

* fix

---------
